woo, float output works

+++ reference_code/halide_hog/test.cpp
@@ -16,8 +16,8 @@ extern "C" {
 #include "gradient.h"
 }
 
-Image<uint8_t> gradient_tester(Image<uint8_t> in) {
-    Image<uint8_t> out(in.width(), in.height(), in.channels()); //TODO: make this a float. 
+Image<float> gradient_tester(Image<uint8_t> in) {
+    Image<float> out(in.width(), in.height(), in.channels()); //TODO: make this a float. 
 
     // Call it once to initialize the halide runtime stuff
     gradient(in, out);
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
     std::string imgName = "../../images_640x480/carsgraz_001.image.png"; //TODO: get jpg support 
     Image<uint8_t> input = load<uint8_t>(imgName.c_str()); //only supports png and ppm
 
-    Image<uint8_t> gradient_result = gradient_tester(input);
+    Image<float> gradient_result = gradient_tester(input);
     double halide_time = (t2.tv_sec - t1.tv_sec) + (t2.tv_usec - t1.tv_usec) / 1000000.0f;
 
     printf("times: %f sec\n", halide_time);
